backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-ai==0.0.14
python-dotenv==1.0.0
httpx>=0.27.2
pydantic>=2.10